Move RoutingDataCache to BasicDataCache as a sharable component
